backend/requirements.txt
django==5.1.7
psycopg2-binary>=2.9.1
djangorestframework>=3.15.0
django-cors-headers>=4.5.0
numpy>=1.26.0
Pillow>=10.4.0
opencv-python>=4.11.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
ultralytics==8.3.89
python-dotenv>=1.1.0
requests>=2.32.0
onvif-zeep